Change container user to root
The main use case of this container is to volume mount files and
directories into the container and use kubectl on them. As such the user
that runs in the container needs to have authority to at least read
those files.

Since docker doesn't include a scheme for UID mapping when volume
mounting, it gets rather tricky to align the UID in the container with
the host, such that the container user can always operate on mounted
files.  Especially generically and since we publish this project to
dockerhub, we can't make assumptions regarding what specific UID happens
to own the mounted files in various organizations. As such we'll just
revert to root until docker inc solves the UID mapping problem in their
product.