Bump SDK to 129.0.0 and pin version
Elimination of non-determinism as to which version of the SDK one ends
with for a given revision of the project. By explicitly specifying the
SDK version in the docker file, the same SDK should be installed
regardless of point-in-time and location the docker image was built.

Also disabled the update nagging and usage statistics, to lessen output
noise when using the SDK, as well as reducing number of unsolicited
outbound network calls.